Small fixes for task 2: fixed janky display of second panel

+++ filesystem/task_2/panel.h
@@ -12,7 +12,7 @@ typedef struct {
   WINDOW *win;
   char *loc;
   struct dirent **content;
-  int content_n;
+  size_t content_n;
   size_t entry_i;
 } mc_panel;
